requirements.txt
python-dotenv==0.9.1
requests==2.28.2
python-telegram-bot==13.0
